requirements.txt
python-slugify
requests
selenium
colorama
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib